--- a/src/join.rs
+++ b/src/join.rs
@@ -54,6 +54,7 @@ bitset_and!{A, B, C, D, E, F, G, H, I, J, K, L, M, N}
 bitset_and!{A, B, C, D, E, F, G, H, I, J, K, L, M, N, O}
 bitset_and!{A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P}
 
+
 /// The only purpose of the `Open` trait is to provide a way
 /// to access the `open` or `open_mut` trait in a generic way
 /// This way the fact that the type is immutable or mutable
@@ -66,51 +67,6 @@ pub trait Open {
     unsafe fn get(Self::Value, Index) -> Self::Type;
 }
 
-/*impl<'a, T> Open for &'a std::sync::RwLockReadGuard<'a, T>
-    where T: Storage
-{
-    type Value = GetRef<'a, T::UnprotectedStorage>;
-    type Mask = &'a BitSet;
-    fn open(self) -> (Self::Mask, Self::Value) {
-        let (l, r) = (**self).open();
-        (l, GetRef(r))
-    }
-}
-
-impl<'b, 'a:'b, T> Open for &'b mut std::sync::RwLockWriteGuard<'a, T>
-    where T: Storage
-{
-    type Value = GetMut<'b, T::UnprotectedStorage>;
-    type Mask = &'b BitSet;
-    fn open(self) -> (Self::Mask, Self::Value) {
-        let (l, r) = (**self).open_mut();
-        (l, GetMut(r))
-    }
-}*/
-
-/*pub struct GetRef<'a, T: 'a>(&'a T);
-impl<'a, T: Component> Get for GetRef<'a, T::Storage> {
-    type Value = &'a T;
-    unsafe fn get(&self, idx: Index) -> Self::Value {
-        self.0.get(idx)
-    }
-}
-
-pub struct GetMut<'a, T: 'a>(&'a mut T);
-impl<'a, T: Component> Get for GetMut<'a, T::Storage> {
-    type Value = &'a mut T;
-    #[allow(mutable_transmutes)]
-    unsafe fn get(&self, idx: Index) -> Self::Value {
-        // This is obviously unsafe and is one of the reasons this
-        // trait is marked as unsafe to being with. It is safe
-        // an an external api point of view because the bitmask
-        // iterator never visits the same index twice, otherwise
-        // this would provide multiple aliased mutable pointers which
-        // is illegal in rust.
-        let x: &mut Self = std::mem::transmute(self);
-        x.0.get_mut(idx)
-    }
-}*/
 
 /// Join is an Iterator over a group of `Storages`
 pub struct Join<O: Open> {
@@ -132,12 +88,8 @@ impl<O: Open> std::iter::Iterator for Join<O> {
     type Item = O::Type;
     fn next(&mut self) -> Option<O::Type> {
         self.keys.next().map(|idx| unsafe {
-            // This is obviously unsafe and is one of the reasons this
-            // trait is marked as unsafe to being with. It is safe
-            // an an external api point of view because the bitmask
-            // iterator never visits the same index twice, otherwise
-            // this would provide multiple aliased mutable pointers which
-            // is illegal in rust.
+            // We only transmute and copy during iteration, which is safe and serves
+            // as a poor man's replacement for the missing re-borrowing semantic.
             let values: O::Value = std::mem::transmute_copy(&self.values);
             O::get(values, idx)
         })
@@ -170,30 +122,6 @@ macro_rules! define_open {
                 ($($from::get($from, i),)*)
             }
         }
-
-        /*impl<'a, $($value,)*, $($from,)*> Get<($($value),*,)> for ($($from),*,)
-            where $($from: Get<$value>),*,
-        {
-            #[allow(non_snake_case)]
-            unsafe fn get(&self, idx: Index) -> ($($value),*,) {
-                let &($(ref $from,)*) = self;
-                ($($from.get(idx)),*,)
-            }
-        }*/
-
-        /*impl<'a, $($from,)*> Join for ($($from),*,)
-            where $($from: Open),*,
-                  ($(<$from as Open>::Mask),*,): BitAnd,
-        {
-            type Mask = <($($from::Mask),*,) as BitAnd>::Value;
-            type Types = ($($from::Type),*,);
-            type Values = ($($from::Value),*,);
-
-            fn join(self) -> Joined<Self::Mask, Self::Types, Self::Values> {
-                let (mask, value) = self.open();
-                Joined::new(mask, value)
-            }
-        }*/
     }
 }