Fixes #32 

Major changes:
- `BitSet` is exposed directly
- `VecStorage` and `HashMapStorage` implementations are much simpler, all the shared logic is moved up to the `MaskedStorage` and `Storage` types
- `Storage` is no longer a trait but a struct, which allows using its methods without any `use` items
- Storages no longer track the generations internally, but instead use the shared generations from the `World`
- `join` module is cleaned up by half

Performance:

| Benchmark  | specs-0.5.1 | specs-master | specs-pr |
| --------------- | --------------- | ----------------- | ----------- |
| pos_vel build | 320 us/iter (+/- 8) | 2,013 us/iter (+/- 139) |1,940 us/iter (+/- 75) |
| pos_vel update | 42 us/iter (+/- 8) | 42 us/iter (+/- 9) |42 us/iter (+/- 11) |
| parallel build | 472 us/iter (+/- 35) | 2,195 us/iter (+/- 89) |2,081 us/iter (+/- 101) |
| parallel update | 66 us/iter (+/- 11) | 64 us/iter (+/- 16) |64 us/iter (+/- 15) |