@kvark 2016-04-24
Note: currently, `Join` is built on top of the `Storage` types returned by the `World`. This means that we have multiple read locks of the G-vec (generations vector), each of which is checked for every component access.

Optimally, the join iterator should only have a single G-vec lock, and check it once of each entity. This should be possible to achieve by integrating the `Join` macro implementation with the `Storage` internals. This should speed it up a little.

TODO: move into a separate issue