@kvark 2016-04-25
@csherratt Have a look at the benchmark table I added. Other than that, this PR is ready to go in. You have the power to hit that green button, I'll much appreciate that ;)